cmake: update the cmake to allow easier integration as a package
So, update the cmake configuration to support the changes for
a 3.x variations to allow the project to be added with find_package
or an add_subdirectory.

This adds in the extract googletest and compile that for the tests
along with a conversion to the target_ usage of cmake for a modern
bit of work.

Signed-off-by: Charles Hardin <charles.hardin@chargepoint.com>